Rename missed package from #279, full.decorators.broken.nodelegateinjectionpoint
